+++ ydb/core/persqueue/pqrb/partition_scale_manager_graph_cmp.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ydb/core/persqueue/utils.h>
+#include <ydb/core/persqueue/public/utils.h>
 #include <ydb/public/sdk/cpp/include/ydb-cpp-sdk/client/topic/control_plane.h>
 #include <span>
 #include <vector>
