+++ ydb/core/persqueue/public/utils.cpp
@@ -1,13 +1,11 @@
 #include "utils.h"
-#include "user_info.h"
 
 #include <ydb/core/base/appdata.h>
-
-#include <deque>
-
+#include <ydb/library/yverify_stream/yverify_stream.h>
 #include <util/generic/algorithm.h>
 #include <util/string/builder.h>
-#include <ydb/library/yverify_stream/yverify_stream.h>
+
+#include <deque>
 
 namespace NKikimr::NPQ {
