+++ ydb/core/persqueue/public/ya.make
@@ -2,9 +2,11 @@ LIBRARY()
 
 SRCS(
     pq_database.cpp
+    utils.cpp
 )
 
 PEERDIR(
+    ydb/core/base
     ydb/core/protos
 )
