+++ ydb/core/persqueue/ut/partitiongraph_ut.cpp
@@ -1,5 +1,5 @@
 
-#include <ydb/core/persqueue/utils.h>
+#include <ydb/core/persqueue/public/utils.h>
 
 #include <library/cpp/testing/unittest/registar.h>
