+++ ydb/core/persqueue/ut/utils_ut.cpp
@@ -1,4 +1,4 @@
-#include <ydb/core/persqueue/utils.h>
+#include <ydb/core/persqueue/public/utils.h>
 
 #include <library/cpp/testing/unittest/registar.h>
