+++ ydb/core/tx/schemeshard/schemeshard__operation_alter_pq.cpp
@@ -9,7 +9,7 @@
 #include <ydb/core/persqueue/partition_key_range/partition_key_range.h>
 #include <ydb/core/persqueue/partition_key_range/partition_key_range_sequence.h>
 #include <ydb/core/persqueue/partition_index_generator/partition_index_generator.h>
-#include <ydb/core/persqueue/utils.h>
+#include <ydb/core/persqueue/public/utils.h>
 
 #include <ydb/services/lib/sharding/sharding.h>
