+++ ydb/services/persqueue_v1/actors/schema_actors.cpp
@@ -3,7 +3,7 @@
 #include "persqueue_utils.h"
 
 #include <ydb/core/client/server/ic_nodes_cache_service.h>
-#include <ydb/core/persqueue/utils.h>
+#include <ydb/core/persqueue/public/utils.h>
 #include <ydb/core/ydb_convert/topic_description.h>
 #include <ydb/core/ydb_convert/ydb_convert.h>
 #include <ydb/public/sdk/cpp/src/library/persqueue/obfuscate/obfuscate.h>
